requirements.txt
llama-cpp-python>=0.2.90
huggingface-hub>=0.20.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
langgraph>=0.0.10
pydantic>=2.0.0
chromadb>=0.4.22
sentence-transformers>=2.3.0
fastapi>=0.109.0
uvicorn>=0.27.0
requests>=2.31.0
nicegui>=1.4.0
pyautogui>=0.9.54
opencv-python>=4.9.0
pillow>=10.0.0
duckduckgo-search>=4.0.0
sounddevice>=0.4.6
faster-whisper>=1.0.0
torch>=2.0.0
numpy>=1.24.0
nvidia-cudnn-cu12

google-api-python-client
google-auth-httplib2
google-auth-oauthlib
diffusers
transformers
accelerate
imageio[ffmpeg]